requirements.txt
Pillow == 8.2.0
beautifulsoup4 == 4.9.3
fake_useragent == 0.1.13
numpy == 1.22.4
requests == 2.25.1
tenacity == 8.1.0
tqdm == 4.59.0
